requirements.txt
asgiref==3.2.4
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
Django==3.0.4
django-bootstrap3==12.0.3
django-leaflet==0.26.0
Faker==0.9.1
hypothesis==5.6.0
idna==2.9
importlib-metadata==1.5.0
mixer==6.1.3
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-sugar==0.9.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sortedcontainers==2.1.0
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.2
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
zipp==3.1.0
